@@ -27,7 +27,6 @@ setup(
         'Programming Language :: Python',
         'Programming Language :: Python :: 3',
         'Programming Language :: Python :: 3 :: Only',
-        'Programming Language :: Python :: 3.10',
         'Programming Language :: Python :: 3.11',
         'Programming Language :: Python :: 3.12',
     ],
@@ -46,5 +45,5 @@ setup(
         'Tracker': 'https://github.com/uoftcprg/pokerkit/issues',
     },
     packages=['pokerkit'],
-    python_requires='>=3.10',
+    python_requires='>=3.11',
 )
